Add an incremental backup mode
The incremental backup strategy slightly reworks the way Medusa currently manages backups. Instead of having one directory per backup to store the SSTables, it has a single directory where only newly added SSTables will be uploaded.
Each backup has a separate manifest file which links to the active sstables at the time of the backup. A single copy of each SSTable would exist at anytime in the storage service and would remain there as long as a manifest references it.
The caching feature of the current backup strategy for GCS was tweaked as follows: instead of replacing the pre-existing files path with their current location in GCS, they are be removed from the list of files to transfer and reintroduced in the manifest after the transfer was completed.
A new file will be placed in the meta folder in order to distinguish incremental backups from full ones : meta/incremental
This is required so that the data path on the NodeBackup object can be adjusted to match both storage patterns.
In case the latest backup is a full one, a subsequent incremental backup will use the “cache” feature of full backups and copy previously backed up files from the backup folder to the data folder (instead of just skipping the files and linking them in the manifest).