Add retrying when doing a backup
The retries are placed deeper in the call trees because retrying the full backup
is too coarse. Most importantly, we don't want to re-upload files that have
already been uploaded succesfully.

Note there is no retrieso medusa.storage.gsutil cp method. This is because
the 'gsutil cp' retries objects under the hood.