[CAS-199] use concurrent.futures.map function
This commit builds off of the changes in CASS-197 with some important changes.
We now use the map function to asynchronously perform the uploads. The function
that performs the upload returns a ManifestObject.

Note that there is no longer a need for a queue.

There is some locking that is needed. Since libcloud is not thread safe, each
thread needs a separate connection. I decided to lazily initialize connections
because we do not know up front how many threads will be used, only the max
number of threads that can be used.